feat(config): add onChange handler to TextField
This commit is concerned with how changes to the
TextField value are tracked, and later used,
within the ConfigScreen component. The TextField
value corresponds with the API key entered by the
user, which is saved as an installation parameter
and used throughout the Contentful space.

The first change erases the previous setAPIKey
class method, which extracted and saved the value
in a non-idiomatic way. The second change adds a
handler method (`handleChange`) to the TextField
component, which overwrites the value of
`imgixAPIKey` in state. The value of `imgixAPIKey`
in the application installation parameters will
only be overwritten once the user presses the
"Save" button.