feat(config): sets a validation message on the TextField if an empty key is passed
The component should signal to users that a non-empty key is required before verifying. This commit will conditionally show a validation message when a user presses `Verify` with an empty field input. The validation is removed once another attempt is made to verify a key, this time with a non-empty value instead.