style: change dialog grid row template to be fixed
This commit makes the grid row sizes be of a fixed size. This addresses
an issue where the image grid rows could not be cut off after a certain
size. Now, we can render an flexible number of placeholder images
without creating an infinite number of rows. Instead, only as many rows
as can fit inside the 460px defined in the row template will be created.

In practice, this means that we can at most have 3 rows of images but
and infinite number of columns.

Future work here would be to make the number of rows more flexible.

Rather than 460px, ideally the row that contains the grid of images
should resize to fit a percentage of the iframe. But I had difficulty
getting the template to base x% off of the available space in the frame
vs the entire window.