style: change the grid margins
Instead of putting 64px margin on the container, this commit puts a 32px
margin on each element within the container. This allowed for better
gutter spacing on the image grid. Before, the right margin on the grid
was far too large and kept the grid from generating an additional column

This commit also sets overflow:hidden for the image grid. This is to
keep the image grid for displaying more rows than there is space on its
assigned grid row.

Future work here could be to make the margins more flexible, ie a % val.
The issue I ran into is that the stylesheet keeps referring to the
window instead of the frame for %, and inconsistently so. This made it
hard to control the margins.