refactor: move gallery image into its own component
This commit refactors the Gallery component so that gallery images are
controlled by their own GalleryImage component. This is mean to reduce
the amount of logic in the Gallery component and simplify the applying
of styles to the gallery images themselves.

The commit also re-organizing imports so that external, internal, and
stylesheet modules are chunked together.