feat(dialog): create a handler fn to return totalImageCount from the Gallery component
Here we employ a pattern of passing a handler function from the parent (Dialog) to the child component (Gallery) which passes back the total number of assets returned for the selected Source. The intention here is to keep the Gallery component modular in that it only handles displaying the images, while the Dialog component will handle all pagination functions.