refactor: image grid
This commit refactors the ImageGrid component so that, rather than
creating an array of images, the GridImage component only handles the
logic relating to a signle image. The array of images is instead created
in the ImageGallery.

This change allows for more fine tuned control of a grid image's hover
and selected state styling.