refactor: create FieldImagePreview component
The FieldImagePreview component is responsible for handling the UI and flow that a user sees on the Field component when an image has been selected from the Dialog. Eventually, users will also be able to clear a current selection, or re-open the Dialog in order to select a replacement image.