feat(field): creates and passes an updateHeight fn to each of the child components
The `FieldExtensionSDK` defines a function `window.updateHeight` that allows the containing iframe's height to be set. We will use this function to set the height based on which component is currently being displayed. This can be used, for example, to give the FieldImagePreview component ample space to display all of its containing elements.