chore: create note component
This commit creates the `Note.tsx` functional component. It has three
arguments, `error`, `type`, and `resetErrorBoundary`. The `error` is an
IxError instance. The `type` is 'primary' | 'positive' | 'negative' |
'warning'. And the `resetErrorBoundary` is a function that is user
defind and called whenever the close button on the note is clicked. In
theory, it should be used to clear the error associated with the note
from the parent component's state.

This component parses the `error.message` value to pull out a link href
and title.

The error messages are split on `$` and then again on `|`.
The `$` is used to parse the message's "link" url. The`|` is used to
parse the message's "link" title. If the `message` string has no `|`
then the `$` is used to parse the message's "link" url and title.

E.g.
`Return to $${APP_CONFIG_URL}|App Config Page$ to enter a valid API key`
Will be parsed as:
`Return to https://app.contentful.com/deeplink?link=apps to enter a valid
API key`.