feat: use note and error helper to throw warnings
This commit uses the new `Note.tsx` component and the `errors.ts` helper
to display a warning || error fallback UI whenever an IxError occurs.

To do this, the commit creates an error slice of state that gets updated
whenever: the API key is invalid, the account has no sources, or the
source has no images. Each displayed error is removed from state so that
only one error can ever be displayed at one time (the oldest error).

This commit also adds another slice of state, `verified`, that comes from
values being written to the contentful SDK. This makes it easier to know
whether or not the API key is valid in the `componentDidMount` lifecycle
hook.