fix: only attempt to parse url when present
This commit fixes a bug where Error message that did not contain a link
where throwing an error, since there was no `|` separator to `.split()`
on. The component now returns an array of two empty strings as a
fallback for an error with only a message string and no link.