fix(config): verify api key on configure
This commit makes it so that the `onConfigure` function calls the
`verifyAPIKey()` function. This way, when a user clicks `save` in the
app configure screen, the validity of the API key is also saved to
state. This avoids failing to save the validity of an API key that
replaces an exisitng and or first install key.