refactor: debounce pagination fn passed from props
The initial approach to debouncing pagination requests was to create a new function that debounced the button onClick handlers. While this approach worked, it came with some strange side effects - e.g. not being able to lead debounce requests.
By refactoring where we are debouncing, in this case the function that handles page state changes, we get an improved experienced and arguably more readable code.