feat: add function to get all content types with compatible fields
The intention of this function is to query the Contentful space API
in order to construct an array of content types with compatible fields
(ie `Object`) that users can elect to add the imgix app to. From this,
we will eventually build the interface that allows users to opt-in
certain fields to use imgix-contentful.