feat: save new target state based on user's selections
Once the user has selected any number of fields to install the app to, we need to construct an object in the shape expected by `onConfigure` to persist these choices. For any content type if there is at least one `enabled` entry then we will create an array of fieldId's corresponding to the enabled (or selected) fields. Once this is returned as `targetState` within `onConfigure`, then the imgix app will be installed in the selected fields.