docs: separately display videos in both md and html
GitHub cannot render <video> tags while the imgix static site cannot render them as markdown. The (not ideal) solution to this may be to include both, but to wrap the markdown videos in `<!-- ix-docs-ignore -->` tags to exclude them from getting rendered on the static site.
Things to look out for here: npmjs.com might render neither or both of these formats. Will need to test it live in order to confirm.