refactor: pull up asset parsing from ImageGallery to Dialog component
In order to implement search, we will need the search UI (text input and button) to have access to the same logic where assets are requested and parsed. Unfortunately there did not appear to be a very react-idiomatic way to handle this situation while also keeping the Dialog and ImageGallery components fully separated.
The current solution here is to pull up all asset-related logic to the Dialog level. Once all processing is done, that data can be passed down to ImageGallery as a prrop for rendering