fix: reset `currentPageIndex` when executing search query
Because we cannot guarantee that a specific search will return N or more pages, we will reset the current page to the first index (0). This way, if a user is currently viewing page 2, the app will not attempt to show them page 2 of their search when one does not exist.