fix: avoid an extra rerender when searching
The `Dialog` component will watch for one of two state changes and rerender if either occurs: if a new Source is selected or if a new page of Assets is requested. Because we reset the `currentPageIndex` when executing a search request, this will trigger an extra request for assets in parallel with the search request. Aside from the fact that it's a superfluous request, it will also trigger a ratelimit error from the imgix API.

Although not ideal, the speediest workaround here is to track if a search is conducted through a variable in state, `isSearching`. Based on if this flag is true, we can tell the component not to rerender. This is not an elegant solution because the `componentDidUpdate` lifecycle listens for changes that will trigger new assets to get requested, but has no visibility if a page change is caused via the pagination UI or the search function. In the future, we should rework the logic here so that we don't need to rely so much on state changes to make the correct request.