feat: add fn to stringify JSON fields within attributes field
Because the keys of nested JSON fields (e.g. `tags`, `custom_fields`,
    `colors`) can vary from one object to the next, this can cause
brittleness for users who are interfacing with the Contentful API using
GraphQL. This is because GraphQL requires users to be explicit about
what fields they want to return, rather than being able to query an
object for all fields regardless of what they are.

To help with this, some users have suggested
[[1](gatsbyjs/gatsby#2490)] [[2](https://forums.fauna.com/t/how-to-store-arbitrary-json-object-via-graphql/142/3)] stringifying nested JSON fields and then parsing it clientside. This better ensures that queries won't break even if the returning data's underlying shape somehow breaks. Now each of the following fields will be returned as a string:

- tags
- custom_fields
- colors.dominant_colors