feat: add replacer fn for stringifying null values
Any fields that do not have values were automatically written as the string "null". This would make it difficult for users to parse/type check this clientside as "null" cannot be parsed or typed checked other than testing that it literally equals the string "null".

Instead, we can replace null values with an empty string which is naturally falsy and provides a little better flexibility to users