feat: resize modal when assets re-render
Before this commit, the AM modal was a fixed height. After this commit,
the AM modal will automatically resize itself to match the asset grid
size. This means that for smaller grids, the modal only takes up
available space; and for larger grids, the modal grows to fit the grid.