fix: loading indicator when no assets present in source
Before this commit, sometimes the loading indicator would stick around
even though loading was done. This was because we were not checking if
the previously selected source and current source ID matched. This meant
that, when they did not, we still treated a source select change as an
image replacement.